+++ indra/llcorehttp/bufferarray.cpp
@@ -304,7 +304,7 @@ int BufferArray::findBlock(size_t pos, size_t * ret_offset)
 
 bool BufferArray::getBlockStartEnd(int block, const char ** start, const char ** end)
 {
-	if (block < 0 || block >= mBlocks.size())
+	if (block < 0 || block >= (int)mBlocks.size())
 	{
 		return false;
 	}
